requirements.txt
flask == 3.1.1
flask-cors == 6.0.0
openai == 1.92.2
python-dotenv == 1.1.0
gunicorn